requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
huggingface-hub>=0.16.0
torch>=2.0.0
